+++ Makefile
@@ -1,0 +1,2 @@
+w-tui:
+	cargo run -- -p innovatetech-development
